WIP: Separating into different files + QR Code Scanning
